@@ -1,4 +1,4 @@
-import Vue from "vue";
+import { createApp } from "vue";
 import App from "./App.vue";
 import router from "./router";
 import "./index.css";
@@ -9,9 +9,7 @@ import "./css/animate.css";
 import "./css/responsive.css";
 import "./css/custom.css";
 
-Vue.config.productionTip = false;
-
-new Vue({
-  router,
-  render: (h) => h(App),
-}).$mount("#app");
+createApp(App)
+  .use()
+  .use(router)
+  .mount("#app");
