@@ -6,232 +6,105 @@
       title="Changa Medical Communities"
     />
     <section id="content">
-      <div class="content-wrap pt-0">
-        <div class="section nomargin med-communities-about">
-          <div class="container">
-            <div class="heading-block center">
-              <h2>About</h2>
-            </div>
-            <div class="row justify-content-center">
-              <div class="col-md-10">
-                <p class="lead-paragraph text-center">
-                  This is a community of M-Changa selected medical
-                  fundraisers.<br />Learn more at:
-                  http://changa.co.ke/community/
-                </p>
-              </div>
-            </div>
-            <div class="row med-com-contact">
-              <div class="col-md-6">
-                <div class="med-contact-inner">
-                  <img class="mx-auto" src="@/assets/images/ic-phone-2.png" />
-                  <div class="contact-descripiton">
-                    Contact Changa Medical Coordinator :
-                  </div>
-                  <div class="contact-detail">254722000000</div>
-                </div>
-              </div>
-              <div class="col-md-6">
-                <div class="med-contact-inner">
-                  <img class="mx-auto" src="@/assets/images/ic-mail-2.png" />
-                  <div class="contact-descripiton">
-                    To request to join Changa Medical Community :
-                  </div>
-                  <div class="contact-detail">support@changa.co.ke</div>
-                </div>
-              </div>
-            </div>
-            <div class="row">
-              <div class="col-md-4 col-4">
-                <div class="statistic-item center">
-                  <div class="counter">$1,906</div>
-                  <p>Funds Raised</p>
-                </div>
-              </div>
-              <div class="col-md-4 col-4">
-                <div class="statistic-item center">
-                  <div class="counter">225</div>
-                  <p>Supporters</p>
-                </div>
-              </div>
-              <div class="col-md-4 col-4">
-                <div class="statistic-item center">
-                  <div class="counter">26</div>
-                  <p>Days Left</p>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-        <div class="section nobg nomargin medical-stories-section">
-          <div class="container">
-            <div class="row">
-              <div class="col-md-4">
-                <div class="medical-story-item">
-                  <img src="@/assets/images/baby-nathael.jpg" />
-                  <div class="medical-raised">Raised : $4,178</div>
-                  <h4 class="medical-title">
-                    Baby Nathaniel Hinga’s Medical Fund
-                  </h4>
-                  <a href="#" class="button mx-auto">Read Their Story</a>
-                </div>
-              </div>
-              <div class="col-md-4">
-                <div class="medical-story-item">
-                  <img src="@/assets/images/baby-nathael.jpg" />
-                  <div class="medical-raised">Raised : $4,178</div>
-                  <h4 class="medical-title">
-                    Baby Nathaniel Hinga’s Medical Fund
-                  </h4>
-                  <a href="#" class="button mx-auto">Read Their Story</a>
-                </div>
-              </div>
-              <div class="col-md-4">
-                <div class="medical-story-item">
-                  <img src="@/assets/images/baby-nathael.jpg" />
-                  <div class="medical-raised">Raised : $4,178</div>
-                  <h4 class="medical-title">
-                    Baby Nathaniel Hinga’s Medical Fund
-                  </h4>
-                  <a href="#" class="button mx-auto">Read Their Story</a>
-                </div>
-              </div>
-              <div class="col-md-4">
-                <div class="medical-story-item">
-                  <img src="@/assets/images/baby-nathael.jpg" />
-                  <div class="medical-raised">Raised : $4,178</div>
-                  <h4 class="medical-title">
-                    Baby Nathaniel Hinga’s Medical Fund
-                  </h4>
-                  <a href="#" class="button mx-auto">Read Their Story</a>
-                </div>
-              </div>
-              <div class="col-md-4">
-                <div class="medical-story-item">
-                  <img src="@/assets/images/baby-nathael.jpg" />
-                  <div class="medical-raised">Raised : $4,178</div>
-                  <h4 class="medical-title">
-                    Baby Nathaniel Hinga’s Medical Fund
-                  </h4>
-                  <a href="#" class="button mx-auto">Read Their Story</a>
-                </div>
-              </div>
-              <div class="col-md-4">
-                <div class="medical-story-item">
-                  <img src="@/assets/images/baby-nathael.jpg" />
-                  <div class="medical-raised">Raised : $4,178</div>
-                  <h4 class="medical-title">
-                    Baby Nathaniel Hinga’s Medical Fund
-                  </h4>
-                  <a href="#" class="button mx-auto">Read Their Story</a>
-                </div>
-              </div>
-              <div class="col-md-4">
-                <div class="medical-story-item">
-                  <img src="@/assets/images/baby-nathael.jpg" />
-                  <div class="medical-raised">Raised : $4,178</div>
-                  <h4 class="medical-title">
-                    Baby Nathaniel Hinga’s Medical Fund
-                  </h4>
-                  <a href="#" class="button mx-auto">Read Their Story</a>
-                </div>
-              </div>
-              <div class="col-md-4">
-                <div class="medical-story-item">
-                  <img src="@/assets/images/baby-nathael.jpg" />
-                  <div class="medical-raised">Raised : $4,178</div>
-                  <h4 class="medical-title">
-                    Baby Nathaniel Hinga’s Medical Fund
-                  </h4>
-                  <a href="#" class="button mx-auto">Read Their Story</a>
-                </div>
-              </div>
-              <div class="col-md-4">
-                <div class="medical-story-item">
-                  <img src="@/assets/images/baby-nathael.jpg" />
-                  <div class="medical-raised">Raised : $4,178</div>
-                  <h4 class="medical-title">
-                    Baby Nathaniel Hinga’s Medical Fund
-                  </h4>
-                  <a href="#" class="button mx-auto">Read Their Story</a>
-                </div>
-              </div>
-            </div>
-            <div class="row med-com-pagination">
-              <div class="col-md-12">
-                <ul class="pagination">
-                  <li class="page-item disabled">
-                    <a class="page-link" href="#">&laquo;</a>
-                  </li>
-                  <li class="page-item active">
-                    <a class="page-link" href="#"
-                      >1 <span class="sr-only">(current)</span></a
-                    >
-                  </li>
-                  <li class="page-item"><a class="page-link" href="#">2</a></li>
-                  <li class="page-item"><a class="page-link" href="#">3</a></li>
-                  <li class="page-item"><a class="page-link" href="#">4</a></li>
-                  <li class="page-item"><a class="page-link" href="#">5</a></li>
-                  <li class="page-item">
-                    <a class="page-link" href="#">&raquo;</a>
-                  </li>
-                </ul>
-              </div>
-            </div>
-          </div>
-        </div>
-        <div class="section nomargin med-communities-comment">
-          <div class="container">
-            <div class="heading-block center">
-              <h2>Well Wisher Comments</h2>
-            </div>
-            <div class="comment-plugin-wrap">
-              <!-- DUMMY TEXT, just replace with the plugin -->
-              <div
-                style="text-align: center; height: 300px; line-height: 300px"
-              >
-                Comments Plugins Here
-              </div>
-              <!-- END DUMMY -->
-            </div>
-          </div>
-        </div>
-      </div>
-      <div class="section nobg nomargin home-faq-sec pt-0">
+      <div class="section nomargin med-communities-about">
         <div class="container">
           <div class="heading-block center">
-            <h2>Have a question? Check out our FAQs page or chat with us</h2>
+            <h2>About</h2>
+          </div>
+          <div class="row justify-content-center">
+            <div class="col-md-10">
+              <p class="lead-paragraph text-center">
+                This is a community of M-Changa selected medical fundraisers.<br />Learn
+                more at: http://changa.co.ke/community/
+              </p>
+            </div>
+          </div>
+          <div class="row med-com-contact">
+            <div class="col-md-6" v-for="v in aboutDetail" :key="v">
+              <AboutCard :src="v.src" :text="v.text" :contact="v.contact" />
+            </div>
           </div>
           <div class="row">
-            <div class="col-md-4">
-              <a href="#" class="button button-icon">
-                <img src="@/assets/images/ic-faq-1.png" />
-                <span>Start A Fundraiser</span>
-              </a>
+            <div class="col-md-4 col-4">
+              <div class="statistic-item center">
+                <div class="counter">$1,906</div>
+                <p>Funds Raised</p>
+              </div>
             </div>
-            <div class="col-md-4">
-              <a href="#" class="button button-icon">
-                <img src="@/assets/images/ic-faq-2.png" />
-                <span>Find Out More</span>
-              </a>
+            <div class="col-md-4 col-4">
+              <div class="statistic-item center">
+                <div class="counter">225</div>
+                <p>Supporters</p>
+              </div>
             </div>
-            <div class="col-md-4">
-              <a href="#" class="button button-icon">
-                <img src="@/assets/images/ic-faq-3.png" />
-                <span>Request a call back</span>
-              </a>
+            <div class="col-md-4 col-4">
+              <div class="statistic-item center">
+                <div class="counter">26</div>
+                <p>Days Left</p>
+              </div>
             </div>
           </div>
         </div>
       </div>
+      <div class="section nobg nomargin medical-stories-section">
+        <div class="container">
+          <div class="row">
+            <div class="col-md-4" v-for="n in 3" :key="n">
+              <ProductCard />
+            </div>
+          </div>
+          <div class="row med-com-pagination">
+            <div class="col-md-12">
+              <Pagination />
+            </div>
+          </div>
+        </div>
+      </div>
+      <div class="section nomargin med-communities-comment">
+        <div class="container">
+          <div class="heading-block center">
+            <h2>Well Wisher Comments</h2>
+          </div>
+          <div class="comment-plugin-wrap">
+            <div style="text-align: center; height: 300px; line-height: 300px">
+              Comments Plugins Here
+            </div>
+          </div>
+        </div>
+      </div>
+      <QuestionBottom />
     </section>
   </div>
 </template>
 <script>
 import Breadcrumbs from "../components/Breadcrumbs.vue";
+import QuestionBottom from "../components/QuestionBottom.vue";
+import Pagination from "../components/Pagination.vue";
+import ProductCard from "../components/ProductCard.vue";
+import AboutCard from "../components/AboutCard.vue";
 export default {
   components: {
     Breadcrumbs,
+    QuestionBottom,
+    Pagination,
+    ProductCard,
+    AboutCard,
+  },
+  data() {
+    return {
+      aboutDetail: [
+        {
+          src: require("@/assets/images/ic-phone-2.png"),
+          text: "Contact Changa Medical Coordinator :",
+          contact: "254722000000",
+        },
+        {
+          src: require("@/assets/images/ic-mail-2.png"),
+          text: "To request to join Changa Medical Community :",
+          contact: "support@changa.co.ke",
+        },
+      ],
+    };
   },
 };
 </script>
