@@ -751,7 +751,8 @@
   </div>
 </template>
 <script>
-import Breadcrumbs from "../components/Breadcrumbs.vue";
+import Breadcrumbs from '../components/Breadcrumbs.vue';
+
 export default {
   components: {
     Breadcrumbs,
